Fixed issues timing in sampling, and 'fix' in log
Timing issue in sampling was related to writing to disk on every sample,
which caused some samples (at the end of trials) to not be written. This
was overly paranoid. Writing to disk is now done everytime
`stop_recording` is called.

The 'fix' variable never appeared as True in log files due to an
incorrect string comparison. The comparison is now removed.